auth/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
autopep8==2.0.4
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
Django==4.2.7
django-cors-headers==4.3.0
django-ratelimit==4.1.0
django-uuidfield==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
kombu==5.3.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
